+++ crates/icp/src/lib.rs
@@ -29,6 +29,7 @@ pub mod identity;
 pub mod manifest;
 pub mod network;
 pub mod package;
+pub mod parsers;
 pub mod prelude;
 pub mod project;
 pub mod settings;
@@ -409,6 +410,7 @@ impl MockProjectLoader {
                         gateway: Gateway {
                             host: "localhost".to_string(),
                             port: Port::Fixed(8000),
+                            domains: vec![],
                         },
                         artificial_delay_ms: None,
                         ii: false,
@@ -428,6 +430,7 @@ impl MockProjectLoader {
                         gateway: Gateway {
                             host: "localhost".to_string(),
                             port: Port::Fixed(8001),
+                            domains: vec![],
                         },
                         artificial_delay_ms: None,
                         ii: false,