+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * feat: Init/call arg files now support raw binary without conversion to hex
 * feat!: Remove argument type inference in init/call args in commands and manifest. Args are always assumed Candid, new parameters allow specifying other formats like hex, and alternate parameters are used to specify loading from a file.
+* feat: Network gateway now supports a `domains` key
 * feat: `icp identity export` now takes an `--encrypt` flag to avoid rendering the key in plaintext
 * feat: Optionally split connected networks' `url` into `api-url` and `http-gateway-url`
 * feat: Allow specifying a version of the network launcher to use
@@ -10,11 +11,12 @@
 * feat: Recipes and prebuilt canisters are now cached locally
 * feat: `icp settings autocontainerize true`, always use a docker container for all networks
 * feat: `icp canister migrate-id` - initiate canister ID migration across subnets
-* feat: install proxy canister when starting managed networks with all identities as controllers (or anonymous + default if more than 10 identities)
+* feat: Install proxy canister when starting managed networks with all identities as controllers (or anonymous + default if more than 10 identities)
   * `icp network status` displays the proxy canister principal
 * feat: `icp network status` display more information about networks
 * feat: `icp canister logs` to display the current canister logs
   * use `--follow` to continuously poll for new logs. `--interval <n>` to poll every `n` seconds
+* feat: Support `k`, `m`, `b`, `t` suffixes in `.yaml` files when specifying cycles amounts
 
 # v0.1.0